pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whgot"
version = "0.1.0"
description = "Identify items from photos or text lists and retrieve structured resale data"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Alexander Pico", email = "arpico@gmail.com" },
]
dependencies = [
    "fastapi>=0.115",
    "httpx>=0.27",
    "jinja2>=3.1",
    "ollama>=0.4",
    "pillow>=10.0",
    "pydantic>=2.0",
    "python-multipart>=0.0.9",
    "rich>=13.0",
    "typer>=0.12",
    "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
whgot = "whgot.cli:app"
whgot-web = "whgot.web_app:run"

[tool.hatch.build.targets.wheel]
packages = ["src/whgot"]
include = [
    "/src/whgot/web_templates/**/*",
    "/src/whgot/web_static/**/*",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
